+++ packages/rstream-gestures/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.1"></a>
+## [0.5.1](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-gestures@0.5.0...@thi.ng/rstream-gestures@0.5.1) (2018-08-27)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/rstream-gestures
+
+<a name="0.5.0"></a>
+# [0.5.0](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-gestures@0.4.18...@thi.ng/rstream-gestures@0.5.0) (2018-08-27)
+
+
+### Features
+
+* **rstream-gestures:** add options for local & scaled positions ([ccc40a9](https://github.com/thi-ng/umbrella/commit/ccc40a9))
+
+
+
+
 <a name="0.4.18"></a>
 ## [0.4.18](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-gestures@0.4.17...@thi.ng/rstream-gestures@0.4.18) (2018-08-24)